[BE] Refactor : 결제요청 알림 보내기

+++ backend/account/src/main/java/com/ssafy/account/api/request/message/PaymentNotificationRequest.java
@@ -11,10 +11,10 @@ import java.io.Serializable;
 @NoArgsConstructor
 public class PaymentNotificationRequest implements Serializable {
     Long receiveMemberId;
-    int price;
+    Long price;
     String sender; // 결제처
 
-    public PaymentNotificationRequest(Long receiveMemberId, int price, String sender) {
+    public PaymentNotificationRequest(Long receiveMemberId, Long price, String sender) {
         this.receiveMemberId = receiveMemberId;
         this.price = price;
         this.sender = sender;
